+++ fiddles/angular2-cli/fiddle-0016-PersistentLinkList/src/app/components/components.module.ts
@@ -1,0 +1,12 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { ListComponent } from './list/list.component';
+import { ListEditorComponent } from './list-editor/list-editor.component';
+
+@NgModule({
+  imports: [
+    CommonModule
+  ],
+  declarations: [ListComponent, ListEditorComponent]
+})
+export class ComponentsModule { }