+++ fiddles/angular2-cli/fiddle-0016-PersistentLinkList/src/app/services/local-storage.ts
@@ -1,0 +1,9 @@
+export interface LocalStorage {
+  readonly tables: string[];
+  connect(): boolean;
+  delete(table: string): boolean;
+  pull(table: string):any;
+  push(table: string, value: any): boolean;
+  exists(table: string): boolean;
+  zero(criteria: string): void;
+}