+++ fiddles/angular2-cli/fiddle-0016-PersistentLinkList/src/app/services/state.ts
@@ -1,0 +1,13 @@
+import {Observable} from 'rxjs/Observable';
+import 'rxjs/add/operator/share';
+import {LocalStorage} from 'local-storage';
+
+
+export interface State {
+
+  databaseService: LocalStorage;
+
+  stateEventChange$: Observable<string>;
+
+  stateEvent: string;
+}